requirements.txt
python-dotenv==1.0.0
rich==13.6.0
typer==0.9.0
colorama==0.4.6
openai>=1.3.0
transformers>=4.35.0
huggingface_hub>=0.19.0
tqdm>=4.65.0
torch>=2.0.0
pydantic==1.10.8
langchain==0.0.335
langchain-openai==0.0.2
langchain-community==0.0.13
langchain-core==0.1.10
pydantic
pydantic_core
